python-flask/user_platform/requirements.txt
# install cmd: pip3 install -r requirements.txt

# core
flask
jinja2
PyYAML
requests

# DB
mysql-connector-python

# VMware vSphere
pyvmomi